+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This is another pre-release that adds new features.
 - Add backups management commands.
 - Add `profiles preview` to see the file contents that would be written on
   apply.
+- Add `--all` flag to profiles `enable` and `disable`.
 
 ## v0.5.0 - 2026-03-27